Code Refactor
Refactored AndroidElement class to interface and implemented AndroidElement interface in UiObject2Element class
