+++ app/src/main/java/io/appium/uiautomator2/core/AccessibilityNodeInfoGetter.java
@@ -1,6 +1,7 @@
 package io.appium.uiautomator2.core;
 
 import android.support.test.uiautomator.Configurator;
+import android.support.test.uiautomator.UiObject;
 import android.support.test.uiautomator.UiObject2;
 import android.view.accessibility.AccessibilityNodeInfo;
 
@@ -17,7 +18,13 @@ public abstract class AccessibilityNodeInfoGetter {
     /**
      * Gets the {@link AccessibilityNodeInfo} associated with the given {@link UiObject2}
      */
-    public static AccessibilityNodeInfo fromUiObject(UiObject2 uiObject2) {
-        return (AccessibilityNodeInfo) invoke(method(UiObject2.class, "getAccessibilityNodeInfo"), uiObject2);
+    public static AccessibilityNodeInfo fromUiObject(Object object) {
+        if (object instanceof UiObject2) {
+            return (AccessibilityNodeInfo) invoke(method(UiObject2.class, "getAccessibilityNodeInfo"), object);
+        } else if (object instanceof UiObject) {
+            return (AccessibilityNodeInfo) invoke(method(UiObject.class, "findAccessibilityNodeInfo"), object);
+        } else {
+            throw new RuntimeException("Unknown object type: " + object.getClass().getName());
+        }
     }
 }
