+++ app/src/main/java/io/appium/uiautomator2/handler/FindElement.java
@@ -32,6 +32,7 @@ import io.appium.uiautomator2.utils.Logger;
 import io.appium.uiautomator2.utils.XMLHierarchy;
 
 import static io.appium.uiautomator2.model.internal.CustomUiDevice.getInstance;
+import static io.appium.uiautomator2.utils.Device.getAndroidElement;
 import static io.appium.uiautomator2.utils.Device.getUiDevice;
 
 public class FindElement extends SafeRequestHandler {
@@ -74,12 +75,12 @@ public class FindElement extends SafeRequestHandler {
             By by = new NativeAndroidBySelector().pickFrom(method, selector);
 
             getUiDevice().waitForIdle();
-            UiObject2 element = this.findElement(by);
+            Object element = this.findElement(by);
             if (element == null) {
                 return new AppiumResponse(getSessionId(request), WDStatus.NO_SUCH_ELEMENT, false);
             } else {
                 String id = UUID.randomUUID().toString();
-                AndroidElement androidElement = new AndroidElement(id, element);
+                AndroidElement androidElement = getAndroidElement(id, element);
                 ke.add(androidElement);
                 JSONObject result = new JSONObject();
                 result.put("ELEMENT", id);
@@ -109,7 +110,7 @@ public class FindElement extends SafeRequestHandler {
     /**
      * returns  UiObject2 for an xpath expression
      */
-    private UiObject2 findElement(By by) throws InvalidSelectorException, ElementNotFoundException, ParserConfigurationException, ClassNotFoundException {
+    private Object findElement(By by) throws InvalidSelectorException, ElementNotFoundException, ParserConfigurationException, ClassNotFoundException {
         if (by instanceof ById) {
             return getInstance().findObject(android.support.test.uiautomator.By.res(by.getElementLocator()));
         } else if (by instanceof ByLinkText) {
