+++ app/src/main/java/io/appium/uiautomator2/handler/PressBack.java
@@ -3,9 +3,10 @@ package io.appium.uiautomator2.handler;
 import io.appium.uiautomator2.handler.request.SafeRequestHandler;
 import io.appium.uiautomator2.http.AppiumResponse;
 import io.appium.uiautomator2.http.IHttpRequest;
-import io.appium.uiautomator2.model.AndroidElement;
 import io.appium.uiautomator2.server.WDStatus;
 import io.appium.uiautomator2.utils.Logger;
+
+import static io.appium.uiautomator2.utils.Device.back;
 
 public class PressBack extends SafeRequestHandler {
 
@@ -16,7 +17,7 @@ public class PressBack extends SafeRequestHandler {
     @Override
     public AppiumResponse safeHandle(IHttpRequest request) {
         Logger.info("Go Back");
-        boolean status = AndroidElement.back();
+        boolean status = back() ;
         return new AppiumResponse(getSessionId(request), WDStatus.SUCCESS, status);
     }
 }
