+++ app/src/main/java/io/appium/uiautomator2/handler/ScrollTo.java
@@ -12,10 +12,11 @@ import java.util.regex.Pattern;
 import io.appium.uiautomator2.handler.request.SafeRequestHandler;
 import io.appium.uiautomator2.http.AppiumResponse;
 import io.appium.uiautomator2.http.IHttpRequest;
-import io.appium.uiautomator2.model.AndroidElement;
 import io.appium.uiautomator2.server.WDStatus;
 import io.appium.uiautomator2.utils.Device;
 import io.appium.uiautomator2.utils.Logger;
+
+import static io.appium.uiautomator2.utils.Device.scrollTo;
 
 public class ScrollTo extends SafeRequestHandler {
 
@@ -33,9 +34,10 @@ public class ScrollTo extends SafeRequestHandler {
             // Extracting (\"Radio Group\") text from the String
             // TODO This logic needs to be changed according to the request body from the Driver
             Matcher m = Pattern.compile("\\(\"([^)]+)\"\\)").matcher(uiSelectorString);
-            while (m.find())
+            while (m.find()) {
                 scrollToString = m.group(1);
-            AndroidElement.scrollTo(scrollToString);
+            }
+            scrollTo(scrollToString);
         } catch (JSONException e) {
             Logger.error("Exception while reading JSON: ", e);
             return new AppiumResponse(getSessionId(request), WDStatus.UNKNOWN_ERROR, e);
