+++ app/src/main/java/io/appium/uiautomator2/handler/Wake.java
@@ -5,9 +5,10 @@ import android.os.RemoteException;
 import io.appium.uiautomator2.handler.request.SafeRequestHandler;
 import io.appium.uiautomator2.http.AppiumResponse;
 import io.appium.uiautomator2.http.IHttpRequest;
-import io.appium.uiautomator2.model.AndroidElement;
 import io.appium.uiautomator2.server.WDStatus;
 import io.appium.uiautomator2.utils.Logger;
+
+import static io.appium.uiautomator2.utils.Device.wake;
 
 public class Wake extends SafeRequestHandler {
 
@@ -18,7 +19,7 @@ public class Wake extends SafeRequestHandler {
     @Override
     public AppiumResponse safeHandle(IHttpRequest request) {
         try {
-            AndroidElement.wake();
+            wake();
         } catch (RemoteException e) {
             Logger.error("Error waking up device");
             return new AppiumResponse(getSessionId(request), WDStatus.UNKNOWN_ERROR, e);
