+++ app/src/main/java/io/appium/uiautomator2/utils/Device.java
@@ -1,7 +1,15 @@
 package io.appium.uiautomator2.utils;
 
+import android.os.RemoteException;
 import android.support.test.InstrumentationRegistry;
 import android.support.test.uiautomator.UiDevice;
+import android.support.test.uiautomator.UiObject2;
+import android.support.test.uiautomator.UiObjectNotFoundException;
+import android.support.test.uiautomator.UiScrollable;
+import android.support.test.uiautomator.UiSelector;
+
+import io.appium.uiautomator2.model.AndroidElement;
+import io.appium.uiautomator2.model.UiObject2Element;
 
 public abstract class Device {
     private static UiDevice uiDevice;
@@ -11,5 +19,28 @@ public abstract class Device {
             uiDevice = UiDevice.getInstance(InstrumentationRegistry.getInstrumentation());
         }
         return uiDevice;
+    }
+
+    public static AndroidElement getAndroidElement(String id, Object element) {
+        if (element instanceof UiObject2) {
+            return new UiObject2Element(id, (UiObject2) element);
+        } else {
+            throw new RuntimeException("Unknown Element type: " + element.getClass().getName());
+        }
+    }
+
+    public static void wake() throws RemoteException {
+        uiDevice.wakeUp();
+    }
+
+    public static void scrollTo(String scrollToString) throws UiObjectNotFoundException {
+        // TODO This logic needs to be changed according to the request body from the Driver
+        UiScrollable uiScrollable = new UiScrollable(new UiSelector().scrollable(true).instance(0));
+        uiScrollable.scrollIntoView(new UiSelector().descriptionContains(scrollToString).instance(0));
+        uiScrollable.scrollIntoView(new UiSelector().textContains(scrollToString).instance(0));
+    }
+
+    public static boolean back() {
+        return uiDevice.pressBack();
     }
 }
